+++ src/app/(user)/cdp/[study]/docs/doc-list.tsx
@@ -3,8 +3,8 @@
 import { useState } from 'react';
 
 import { Button } from '@/components/ui/button';
-import { getMissionFiles } from '@/drive/files';
-import { DriveFile, googleUrl } from '@/drive/types';
+import { getMissionFiles } from '@/google/drive/files';
+import { DriveFile, googleUrl } from '@/google/drive/types';
 
 import { FileExplorerBox } from './file-explorer';
